change all type of data to double

+++ src/students/items/Food.java
@@ -1,13 +1,13 @@
 package students.items;
 
 public class Food extends Item {
-	private int age = 0;
-	private int maturationAge;
-	private int deathAge;
-	private int monteryValue;
+	private double age = 0;
+	private double maturationAge;
+	private double deathAge;
+	private double monteryValue;
 	private String itemName = " ";
 
-	public Food(int maturationAge, int deathAge, int monteryValue) {
+	public Food(double maturationAge, double deathAge,double monteryValue) {
 		super(maturationAge, deathAge, monteryValue);
 		this.age = getAge();
 	}
